+++ postgresql-cdc/src/main/scala/akka/stream/alpakka/postgresqlcdc/impl/PostgreSQL.scala
@@ -2,12 +2,13 @@
  * Copyright (C) 2016-2018 Lightbend Inc. <http://www.lightbend.com>
  */
 
-package akka.stream.alpakka.postgresqlcdc
+package akka.stream.alpakka.postgresqlcdc.impl
 
 import java.sql.{Connection, DriverManager, PreparedStatement}
 
 import akka.annotation.InternalApi
 import akka.event.LoggingAdapter
+import akka.stream.alpakka.postgresqlcdc.{Mode, Modes, Plugin}
 
 import scala.collection.mutable.ArrayBuffer
 import scala.util.Try
